w3c-reports/12092014/mac-wk2/runned-tests/workers/interfaces/WorkerUtils/importScripts/006-expected.txt

CONSOLE MESSAGE: line 12: DataCloneError: DOM Exception 25: An object could not be cloned.

Harness Error. harness_status.status = 1 , harness_status.message = DataCloneError: DOM Exception 25: An object could not be cloned.
